Tighten the type signature of DEFINE_auto.
This expresses the fact that the callable in the flag holder has the same overall signature as `fn` (and not just the same return type).

This change also introduces some static pytype assertions into the test. In this case we're checking that Pytype infers the correct types for the flag holder values; see the PyType FAQ [0] for more information.

[0] https://google.github.io/pytype/faq.html

PiperOrigin-RevId: 388633853
Change-Id: I9bc8a881e24e4adc6331a5f1b16f856e99707b9f